// src/generated/api/models/ResponseDtoAccessTokenResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 모든 API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0.4
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { AccessTokenResponse } from './AccessTokenResponse';
import {
    AccessTokenResponseFromJSON,
    AccessTokenResponseFromJSONTyped,
    AccessTokenResponseToJSON,
    AccessTokenResponseToJSONTyped,
} from './AccessTokenResponse';

/**
 * 
 * @export
 * @interface ResponseDtoAccessTokenResponse
 */
export interface ResponseDtoAccessTokenResponse {
    /**
     * 
     * @type {number}
     * @memberof ResponseDtoAccessTokenResponse
     */
    status?: number;
    /**
     * 
     * @type {string}
     * @memberof ResponseDtoAccessTokenResponse
     */
    message?: string;
    /**
     * 
     * @type {AccessTokenResponse}
     * @memberof ResponseDtoAccessTokenResponse
     */
    data?: AccessTokenResponse;
    /**
     * 
     * @type {number}
     * @memberof ResponseDtoAccessTokenResponse
     */
    timestamp?: number;
}

/**
 * Check if a given object implements the ResponseDtoAccessTokenResponse interface.
 */
export function instanceOfResponseDtoAccessTokenResponse(value: object): value is ResponseDtoAccessTokenResponse {
    return true;
}

export function ResponseDtoAccessTokenResponseFromJSON(json: any): ResponseDtoAccessTokenResponse {
    return ResponseDtoAccessTokenResponseFromJSONTyped(json, false);
}

export function ResponseDtoAccessTokenResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ResponseDtoAccessTokenResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'status': json['status'] == null ? undefined : json['status'],
        'message': json['message'] == null ? undefined : json['message'],
        'data': json['data'] == null ? undefined : AccessTokenResponseFromJSON(json['data']),
        'timestamp': json['timestamp'] == null ? undefined : json['timestamp'],
    };
}

export function ResponseDtoAccessTokenResponseToJSON(json: any): ResponseDtoAccessTokenResponse {
    return ResponseDtoAccessTokenResponseToJSONTyped(json, false);
}

export function ResponseDtoAccessTokenResponseToJSONTyped(value?: ResponseDtoAccessTokenResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'status': value['status'],
        'message': value['message'],
        'data': AccessTokenResponseToJSON(value['data']),
        'timestamp': value['timestamp'],
    };
}